Add Makefile.PL command line options -y and -n
It would be nice to have a way to run Makefile.PL without user
interaction if you know in advance whether or not you want to install
List::BinarySearch::XS if it is not already present.

This commit adds command line options -y (install) and -n (do not
install) to Makefile.PL. They are implemented using core module
Getopt::Std. The user who wants a more automated installation can add a
CPAN prefs file to provide the option when Makefile.PL is run.